move command to cmd folder
