+++ cmd/kafeman/completion/collection.go
